+++ packages/experimental/Icon/src/Icon.types.ts
@@ -1,34 +1,31 @@
 import { SvgProps } from 'react-native-svg';
 import { AccessibilityProps, ImageProps, ImageStyle, StyleProp, TextStyle } from 'react-native';
+import { ColorValue } from '@fluentui-react-native/theme-types';
 export const iconName = 'Icon';
 
-export interface FontIconProps
-{
+export interface FontIconProps {
   fontFamily?: string;
   fontSrcFile?: string;
   codepoint: number;
   fontSize?: number;
 }
 
-export interface SvgIconProps
-{
+export interface SvgIconProps {
   uri?: string;
   src?: React.FC<SvgProps>;
   viewBox?: string;
 }
 
-export interface RasterImageIconProps
-{
+export interface RasterImageIconProps {
   src: ImageProps['source'];
 }
 
-export interface IconProps extends AccessibilityProps
-{
+export interface IconProps extends AccessibilityProps {
   fontSource?: FontIconProps;
   svgSource?: SvgIconProps;
   rasterImageSource?: RasterImageIconProps;
   style?: StyleProp<ImageStyle | TextStyle>;
   width?: number;
   height?: number;
-  color?: string;
+  color?: ColorValue;
 }