+++ packages/components/RadioGroup/CHANGELOG.md
@@ -1,8 +1,36 @@
 # Change Log - @fluentui-react-native/radio-group
 
-This log was last generated on Mon, 01 Mar 2021 22:53:12 GMT and should not be manually modified.
+This log was last generated on Wed, 07 Apr 2021 23:41:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.3
+
+Wed, 07 Apr 2021 23:41:02 GMT
+
+### Patches
+
+- Update Radio Button focus border style to be WCAG 2.1 compliant (#653) (tamasane@gmail.com)
+
+## 0.9.2
+
+Wed, 07 Apr 2021 22:19:35 GMT
+
+### Patches
+
+- Update radio button focus border style (krsiler@microsoft.com)
+
+## 0.9.0
+
+Sat, 06 Mar 2021 00:20:05 GMT
+
+### Minor changes
+
+- Update to react-native 0.63 (30809111+acoates-ms@users.noreply.github.com)
+
+### Patches
+
+- Base theme on macOS changed to apple theme (saadnajmi2@gmail.com)
 
 ## 0.8.5