+++ packages/experimental/Avatar/CHANGELOG.md
@@ -1,8 +1,64 @@
 # Change Log - @fluentui-react-native/experimental-avatar
 
-This log was last generated on Mon, 01 Mar 2021 22:53:12 GMT and should not be manually modified.
+This log was last generated on Wed, 07 Apr 2021 23:41:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.1
+
+Wed, 07 Apr 2021 23:41:02 GMT
+
+### Patches
+
+- Add the rainbow gradient border image test case to experimental avatar (#654) (tamasane@gmail.com)
+
+## 0.7.0
+
+Wed, 07 Apr 2021 22:19:35 GMT
+
+### Minor changes
+
+- add test case for custom border image (sanajmi@microsoft.com)
+
+### Patches
+
+- Add an NSNull check to the image property (sanajmi@microsoft.com)
+
+## 0.6.2
+
+Thu, 11 Mar 2021 19:48:16 GMT
+
+### Patches
+
+- Update FluentTester to be mobile friendly on iOS (saadnajmi2@gmail.com)
+
+## 0.6.1
+
+Wed, 10 Mar 2021 04:33:26 GMT
+
+### Patches
+
+- Add a native theming module for iOS (saadnajmi2@gmail.com)
+
+## 0.6.0
+
+Mon, 08 Mar 2021 18:43:59 GMT
+
+### Minor changes
+
+- Update FluentUI Apple dependency to 0.2.2, set minimum iOS to 13 (saadnajmi2@gmail.com)
+
+## 0.5.0
+
+Sat, 06 Mar 2021 00:20:05 GMT
+
+### Minor changes
+
+- Update to react-native 0.63 (30809111+acoates-ms@users.noreply.github.com)
+
+### Patches
+
+- Base theme on macOS changed to apple theme (saadnajmi2@gmail.com)
 
 ## 0.4.3