+++ packages/libraries/core/CHANGELOG.md
@@ -1,8 +1,52 @@
 # Change Log - @fluentui/react-native
 
-This log was last generated on Mon, 01 Mar 2021 22:53:12 GMT and should not be manually modified.
+This log was last generated on Wed, 07 Apr 2021 23:41:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.23.5
+
+Wed, 07 Apr 2021 23:41:02 GMT
+
+### Patches
+
+- upgrade react-native-svg (tamasane@gmail.com)
+
+## 0.23.4
+
+Wed, 07 Apr 2021 22:19:35 GMT
+
+### Patches
+
+- Add repository link (krsiler@microsoft.com)
+
+## 0.23.1
+
+Thu, 11 Mar 2021 19:48:16 GMT
+
+### Patches
+
+- Update FluentTester to be mobile friendly on iOS (saadnajmi2@gmail.com)
+
+## 0.23.0
+
+Wed, 10 Mar 2021 04:33:26 GMT
+
+### Minor changes
+
+- Add a native theming module for iOS (saadnajmi2@gmail.com)
+
+## 0.22.0
+
+Sat, 06 Mar 2021 00:20:05 GMT
+
+### Minor changes
+
+- Update to react-native 0.63 (30809111+acoates-ms@users.noreply.github.com)
+
+### Patches
+
+- Base theme on macOS changed to apple theme (saadnajmi2@gmail.com)
 
 ## 0.21.13