+++ packages/theming/win32-theme/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/win32-theme
 
-This log was last generated on Fri, 05 Feb 2021 08:02:13 GMT and should not be manually modified.
+This log was last generated on Mon, 15 Mar 2021 06:53:17 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.1
+
+Mon, 15 Mar 2021 06:53:17 GMT
+
+### Patches
+
+- Public source files that are needed for source map (30809111+acoates-ms@users.noreply.github.com)
+
+## 0.6.0
+
+Sat, 06 Mar 2021 00:20:05 GMT
+
+### Minor changes
+
+- Update to react-native 0.63 (30809111+acoates-ms@users.noreply.github.com)
 
 ## 0.5.2